cleanup(Compiler): Remove the original (now unused) ReflectorEmitter.
This diff looks large because reflector_2 --> reflector.

PiperOrigin-RevId: 182610890
