+++ angular/lib/src/source_gen/template_compiler/code_builder.dart
@@ -33,7 +33,7 @@ String buildGeneratedCode(
   // Generated code.
   final allocator = new Allocator.simplePrefixing();
   final compilerOutput = outputs.templatesSource?.source ?? '';
-  final reflectableOutput = new ReflectableEmitter.useCodeBuilder(
+  final reflectableOutput = new ReflectableEmitter(
     outputs.reflectableOutput,
     new LibraryReader(element),
     allocator: allocator,
